Migrate some library changes
